Move repr stuff to new file.
